为自动模式下的Note分离updae代码 (#72)
* 整理TapNote的代码

* 按照梦佬的建议修改

* 整理DragNote的代码

* 整理BreakNote的代码

* 为自动模式下的Hold、Click分离updae代码

* 为AutoMode下的Click、Hold添加Log

* 加个空格

+++ Cyan-Stars/Assets/Scripts/Gameplay/GameManager.cs
@@ -193,9 +193,9 @@ namespace CyanStars.Gameplay
             Debug.Log("时间轴创建完毕");
         }
 
-        private int GetFullScore()
+        private float GetFullScore()
         {
-            int fullScore = 0;
+            float fullScore = 0;
             foreach (var layer in TimelineSo.musicTimelineData.LayerDatas)
             {
                 foreach (var clip in layer.ClipDatas)
